pyproject.toml
[project]
name = "gai_zoo"
version = "0.1.0"
dependencies = [
    # --------- pytorch --------- #
    "torch==2.4.1",
    "torchvision",

    # --------- huggingface --------- #
    "diffusers==0.32.2",
    "accelerate==1.0.0",
    "datasets==3.0.1",

    # --------- flow matching --------- #
    "torchdiffeq",
    "flow_matching",

    # --------- loggers --------- #
    "tensorboard",

    # --------- evaluation --------- #
    "thop",

    # --------- others --------- #
    "rootutils",
    "pre-commit",
    "rich",
    "pytest",
    "omegaconf",
    "tqdm",
    "ruff",

    # --------- CLI --------- #
    "fire",
    "hydra-core",

    # --------- data --------- #
    "numpy",

    # --------- visualization --------- #
    "seaborn",
    "scikit-learn",
    "matplotlib",
]

[tool.setuptools]
py-modules = []

# ==================== Ruff Setting ====================
[tool.ruff]
line-length = 80
indent-width = 4
target-version = "py312"

[tool.ruff.lint]
select = ["E", "F", "I", "UP"]
fixable = ["ALL"]

[tool.ruff.pydocstyle]
convention = "google"

[tool.ruff.format]
quote-style = "double"
exclude = [
    "src/train.py",
    "src/eval.py",
]